Fix kodak/dc120 camlib compilation warnings.
* find_folder(...): Change 'folder_len' variable type to match 'strlen' return type.
* dc120_get_file(...): Add 'unsigned int' cast to suppress signed-unsigned comparison warning. This does not change behavior of the code.

+++ camlibs/kodak/dc120/dc120.c
@@ -86,7 +86,7 @@ static int find_folder( Camera *camera, const char *folder,
 {
     CameraList *albums = NULL;
     const char* album_name;
-    int folder_len;
+    size_t folder_len;
     int i;
     char *dc120_folder_card   = _("CompactFlash Card");
